requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
contextlib2==21.6.0
cryptography==45.0.5
Django==5.2.4
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
injector==0.22.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
